requirements.txt
dspy-ai
pydantic
google-generativeai
tqdm
absl-py
anndata
numpy
